refactor(ActiveEditorLockManager): Extract window state management code.
