+++ lib/editor/components/GtfsEditor.js
@@ -21,6 +21,7 @@ import {getConfigProperty} from '../../common/util/config'
 import {getEntityName, getTableById} from '../util/gtfs'
 import {getEditorEnabledState, GTFS_ICONS} from '../util/ui'
 import type {Props as ContainerProps} from '../containers/ActiveGtfsEditor'
+import ActiveEditorLockManager from '../containers/ActiveEditorLockManager'
 import type {
   ControlPoint,
   Coordinates,
@@ -37,14 +38,13 @@ import type {
   EditorStatus,
   EditorTables,
   ManagerUserState,
-  MapState,
-  StatusState
+  MapState
 } from '../../types/reducers'
 import type {EditorValidationIssue} from '../util/validation'
 
 import EntityDetails from './EntityDetails'
-import EditorSidebar from './EditorSidebar'
 import EditorHelpModal from './EditorHelpModal'
+import EditorSidebar from './EditorSidebar'
 import EditorMap from './map/EditorMap'
 
 type Props = ContainerProps & {
@@ -59,7 +59,6 @@ type Props = ContainerProps & {
   addStopAtInterval: typeof stopStrategiesActions.addStopAtIntersection,
   addStopAtPoint: typeof stopStrategiesActions.addStopAtPoint,
   addStopToPattern: typeof stopStrategiesActions.addStopToPattern,
-  checkLockStatus: typeof editorActions.checkLockStatus,
   clearGtfsContent: typeof activeActions.clearGtfsContent,
   cloneGtfsEntity: typeof editorActions.cloneGtfsEntity,
   constructControlPoint: typeof mapActions.constructControlPoint,
@@ -71,7 +70,6 @@ type Props = ContainerProps & {
   editSettings: EditSettingsState,
   entities: Array<Entity>,
   entityEdited: boolean,
-  errorStatus: StatusState,
   feedInfo: FeedInfo,
   feedIsLocked: boolean,
   feedSource: Feed,
@@ -84,7 +82,6 @@ type Props = ContainerProps & {
   hasRoutes: boolean,
   hideTutorial: boolean,
   loadFeedVersionForEditing: typeof snapshotActions.loadFeedVersionForEditing,
-  lockEditorFeedSource: typeof editorActions.lockEditorFeedSource,
   mapState: MapState,
   namespace: string,
   newGtfsEntity: typeof editorActions.newGtfsEntity,
@@ -97,7 +94,6 @@ type Props = ContainerProps & {
   refreshBaseEditorData: typeof activeActions.refreshBaseEditorData,
   removeControlPoint: typeof mapActions.removeControlPoint,
   removeEditorLock: typeof editorActions.removeEditorLock,
-  removeEditorLockLastGasp: typeof editorActions.removeEditorLockLastGasp,
   removeStopFromPattern: typeof stopStrategiesActions.removeStopFromPattern,
   resetActiveGtfsEntity: typeof activeActions.resetActiveGtfsEntity,
   saveActiveGtfsEntity: typeof activeActions.saveActiveGtfsEntity,
@@ -131,13 +127,6 @@ type State = {
   activeTableId: ?string
 }
 
-/**
- * Figures out the correct event name to use.
- */
-function getNormalizedEvent (eventName: string) {
-  return window.attachEvent ? `on${eventName}` : eventName
-}
-
 export default class GtfsEditor extends Component<Props, State> {
   componentWillMount () {
     this.setState({
@@ -150,73 +139,6 @@ export default class GtfsEditor extends Component<Props, State> {
     // a successful snapshot/import when starting from scratch).
     this.props.clearGtfsContent()
     this._refreshEditor()
-  }
-
-  // USE_FOR_GTFSPLUS_LOCK
-  componentCleanUp = () => {
-    // When the user exits the editor, as a last-gasp action, remove the editor lock on the feed.
-    const { feedSourceId, removeEditorLockLastGasp, stopLockTimer } = this.props
-    stopLockTimer()
-    removeEditorLockLastGasp(feedSourceId, editorActions.GTFS_EDITOR_LOCK)
-  }
-
-  // USE_FOR_GTFSPLUS_LOCK
-  onFocus = () => {
-    const { checkLockStatus, feedSourceId, namespace } = this.props
-    if (namespace) {
-      // Only claim (back) an editor lock if an editor session (i.e. namespace) has been created.
-      checkLockStatus(feedSourceId, editorActions.GTFS_EDITOR_LOCK)
-    }
-  }
-
-  // USE_FOR_GTFSPLUS_LOCK
-  onVisibilityChange = () => {
-    const {
-      errorStatus,
-      feedSourceId,
-      lockEditorFeedSource,
-      namespace,
-      stopLockTimer
-    } = this.props
-    if (document.visibilityState === 'visible') {
-      // If the page is visible/activated again, resume lock check-in,
-      // unless a modal prompt is shown or no editor was loaded for this feed.
-      if (!errorStatus.modal && !!namespace) {
-        lockEditorFeedSource(feedSourceId, editorActions.GTFS_EDITOR_LOCK)
-      }
-    } else {
-      // When the user exits the editor (i.e. switches, closes, or reloads the tab/window,
-      // or navigates away using the browser buttons),
-      // stop the editor lock timer (don't remove the lock in case the page gets activated again).
-      // Note: this case does not cover the user navigating to other datatool views using regular links from the ui,
-      //   see componentWillUnmount for that.
-      stopLockTimer()
-    }
-  }
-
-  // USE_FOR_GTFSPLUS_LOCK
-  componentDidMount () {
-    // If the browser window/tab is closed, the component does not have a chance
-    // to run componentWillUnmount. This event listener runs clean up in those
-    // cases.
-    window.addEventListener(getNormalizedEvent('pagehide'), this.componentCleanUp)
-
-    // Listen to the window focus event so we can check for things like editor lock status right away.
-    window.addEventListener(getNormalizedEvent('focus'), this.onFocus)
-
-    // Listen to the page visibilityChange event so we can check for things like editor lock status
-    // or pause/resume lock timers right away.
-    window.addEventListener(getNormalizedEvent('visibilitychange'), this.onVisibilityChange)
-  }
-
-  // USE_FOR_GTFSPLUS_LOCK
-  componentWillUnmount () {
-    // Run component clean-up
-    this.componentCleanUp()
-    // And remove the event handlers for normal unmounting
-    window.removeEventListener(getNormalizedEvent('pagehide'), this.componentCleanUp)
-    window.removeEventListener(getNormalizedEvent('focus'), this.onFocus)
-    window.removeEventListener(getNormalizedEvent('visibilitychange'), this.onVisibilityChange)
   }
 
   componentDidUpdate (prevProps: Props) {
@@ -244,7 +166,6 @@ export default class GtfsEditor extends Component<Props, State> {
   }
 
   // FIXME: needs to be updated for GraphQL
-  // USE_FOR_GTFSPLUS_LOCK
   componentWillReceiveProps (nextProps: Props) {
     const {
       clearGtfsContent,
@@ -359,6 +280,7 @@ export default class GtfsEditor extends Component<Props, State> {
       hideTutorial,
       loadFeedVersionForEditing,
       mapState,
+      namespace,
       routeParams,
       sidebarExpanded,
       status,
@@ -382,6 +304,11 @@ export default class GtfsEditor extends Component<Props, State> {
     return (
       <div className='GtfsEditor'>
         <Title>{this.getEditorTitle()}</Title>
+        <ActiveEditorLockManager
+          disableLock={!namespace}
+          feedSourceId={feedSourceId}
+          itemToLock={editorActions.GTFS_EDITOR_LOCK}
+        />
         <EditorSidebar
           activeComponent={activeComponent}
           editingIsDisabled={editingIsDisabled}
