+++ lib/editor/containers/ActiveGtfsEditor.js
@@ -38,13 +38,10 @@ import {
   updateEditSetting
 } from '../actions/active'
 import {
-  checkLockStatus,
   cloneGtfsEntity,
-  lockEditorFeedSource,
   newGtfsEntity,
   newGtfsEntities,
   removeEditorLock,
-  removeEditorLockLastGasp,
   stopLockTimer,
   uploadBrandingAsset
 } from '../actions/editor'
@@ -90,7 +87,6 @@ const mapStateToProps = (state: AppState, ownProps: Props) => {
   // console.log(activeEntityId, activeEntity, active, active.entity)
   const validationErrors = getValidationErrors(state)
   const user = state.user
-  const errorStatus = state.status
 
   // find the containing project
   const project = findProjectByFeedSource(state.projects.all, feedSourceId)
@@ -114,7 +110,6 @@ const mapStateToProps = (state: AppState, ownProps: Props) => {
     editSettings,
     entities,
     entityEdited,
-    errorStatus,
     feedInfo,
     feedIsLocked,
     feedSource,
@@ -146,7 +141,6 @@ const mapDispatchToProps = {
   addStopAtInterval,
   addStopAtPoint,
   addStopToPattern,
-  checkLockStatus,
   clearGtfsContent,
   cloneGtfsEntity,
   constructControlPoint,
@@ -158,13 +152,11 @@ const mapDispatchToProps = {
   handleControlPointDragEnd,
   handleControlPointDragStart,
   loadFeedVersionForEditing,
-  lockEditorFeedSource,
   newGtfsEntities,
   newGtfsEntity,
   refreshBaseEditorData,
   removeControlPoint,
   removeEditorLock,
-  removeEditorLockLastGasp,
   removeStopFromPattern,
   resetActiveGtfsEntity,
   saveActiveGtfsEntity,
